# contract/src/main/swagger/SchemaRegistry.yaml
...
swagger: "2.0"
info:
  description: "REST APIs for Pravega Schema Registry."
  version: "0.0.1"
  title: Pravega Schema Registry APIs
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0"
basePath: "/v1"
...
schemes:
  - http
paths:
  /groups:
    get:
      tags:
      - "Group"
      operationId: listGroups
      description: List all groups
      produces:
        - application/json
      parameters:
        - in: query
          name: continuationToken
          type: string
          description: Continuation token
        - in: query
          name: limit
          type: integer
          description: The numbers of items to return
      required:
          - limit
      responses:
        200:
          description: List of all groups
          schema:
            $ref: "#/definitions/ListGroupsResponse"
        500:
          description: Internal server error while fetching the list of Groups
    post:
      tags:
      - "Group"
      operationId: createGroup
      description: Create a new Group
      consumes:
        - application/json
      parameters:
        - in: body
          name: CreateGroupRequest
          description: The Group configuration
          required: true
          schema:
            type: object
            properties:
              groupName:
                  type: string
              groupProperties:
                  $ref: "#/definitions/GroupProperties"
            required:
                - groupName
                - groupProperties
      responses:
        201:
          description: Successfully added group
        409:
          description: Group with given name already exists
        500:
          description: Internal server error while creating a Group
  /groups/{groupName}:
    parameters:
      - in: path
        name: groupName
        description: Group name
        required: true
        type: string
    get:
      tags:
      - "Group"
      operationId: getGroupProperties
      description: Fetch the properties of an existing Group
      produces:
        - application/json
      responses:
        200:
          description: Found Group properties
          schema:
            $ref: "#/definitions/GroupProperties"
        404:
          description: Group with given name not found
        500:
          description: Internal server error while fetching Group details
    delete:
      tags:
      - "Group"
      operationId: deleteGroup
      description: Delete a Group
      responses:
        204:
          description: Successfully deleted the Group
        500:
          description: Internal server error while deleting the Group
  /groups/{groupName}/history:
    parameters:
      - in: path
        name: groupName
        description: Group name
        required: true
        type: string
    get:
      tags:
      - "Group"
      operationId: getGroupHistory
      description: Fetch the history of schema evolution of a Group
      produces:
        - application/json
      responses:
        200:
          description: Found Group history
          schema:
            $ref: "#/definitions/GroupHistory"
        404:
          description: Group with given name not found
        500:
          description: Internal server error while fetching Group history
  /groups/{groupName}/rules:
    parameters:
      - in: path
        name: groupName
        description: Group name
        required: true
        type: string
    put:
      tags:
      - "Group"
      operationId: updateSchemaValidationRules
      description: update schema validation rules of an existing Group
      consumes:
        - application/json
      parameters:
        - in: body
          name: UpdateValidationRulesRequest
          description: update group policy
          required: true
          schema:
            type: object
            properties:
              validationRules:
                  $ref: "#/definitions/SchemaValidationRules"
              previousRules:
                  $ref: "#/definitions/SchemaValidationRules"
                  nullable: true
            required:
              - validationRules      
      responses:
        200:
          description: Updated schema validation policy
        404:
          description: Group with given name not found
        409:
          description: Write conflict 
        500:
          description: Internal server error while updating Group's schema validation rules 
  /groups/{groupName}/schemas:
    parameters:
      - in: path
        name: groupName
        description: Group name
        required: true
        type: string
      - in: query
        name: type
        type: string
        description: Type of object
    get:
      tags:
      - "Group"
      operationId: getSchemas
      description: Fetch latest schema versions for all objects identified by SchemaInfo#type under a Group. If query param type is specified then latest schema for the type is returned. 
      produces:
        - application/json
      responses:
        200:
          description: Latest schemas for all objects identified by SchemaInfo#type under the group
          schema:
            $ref: "#/definitions/SchemaVersionsList"
        404:
          description: Group with given name not found
        500:
          description: Internal server error while fetching Group's latest schemas
...
  /groups/{groupName}/schemas/versions:
    parameters:
      - in: path
        name: groupName
        description: Group name
        required: true
        type: string
    post:
      tags:
      - "Group"
      operationId: addSchema
      description: Adds a new schema to the group
      consumes:
        - application/json
      parameters:
        - in: body
          name: schemaInfo
          description: Add new schema to group
          required: true
          schema:
              $ref: "#/definitions/SchemaInfo"
      produces:
        - application/json
      responses:
        201:
          description: Successfully added schema to the group
          schema:
            $ref: "#/definitions/VersionInfo"
        404:
          description: Group not found
        409:
          description: Incompatible schema
        417:
          description: Invalid serialization format
        500:
          description: Internal server error while adding schema to group
  /groups/{groupName}/schemas/versions/find:
    parameters:
      - in: path
        name: groupName
        description: Group name
        required: true
        type: string
    post:
      tags:
      - "Group"
      operationId: getSchemaVersion
      description: Get the version for the schema if it is registered. It does not automatically register the schema. To add new schema use addSchema   
      consumes:
        - application/json
      parameters:
        - in: body
          name: schemaInfo
          description: Get schema corresponding to the version
          required: true
          schema:
              $ref: "#/definitions/SchemaInfo"
      produces:
        - application/json
      responses:
        200:
          description: Schema version
          schema:
            $ref: "#/definitions/VersionInfo"
        404:
          description: Group with given name not found
        500:
          description: Internal server error fetching version for schema
  /groups/{groupName}/schemas/versions/{versionOrdinal}:
    parameters:
      - in: path
        name: groupName
        description: Group name
        required: true
        type: string
      - in: path
        name: versionOrdinal
        description: Version ordinal
        required: true
        type: integer
        format: int32
    get:
      tags:
      - "Group"
      operationId: getSchemaFromVersion
      description: Get schema from the version ordinal that uniquely identifies the schema in the group. 
      produces:
        - application/json
      responses:
        200:
          description: Schema corresponding to the version
          schema:
            $ref: "#/definitions/SchemaInfo"
        404:
          description: Group with given name not found
        500:
          description: Internal server error while fetching schema from version  
    delete:
      tags:
      - "Group"
      operationId: deleteSchemaVersion
      description: Delete schema version from the group.
      produces:
        - application/json
      responses:
        204:
          description: Schema corresponding to the version
        404:
          description: Group with given name not found
        500:
          description: Internal server error while deleting schema from group  
  /groups/{groupName}/schemas/versions/validate:
    parameters:
      - in: path
        name: groupName
        description: Group name
        required: true
        type: string
    post:
      tags:
      - "Group"
      operationId: validate
      description: Checks if given schema is compatible with schemas in the registry for current policy setting. 
      consumes:
        - application/json
      parameters:
        - in: body
          name: ValidateRequest
          description: Checks if schema is valid with respect to supplied validation rules
          required: true
          schema:
            type: object
            properties:
              schemaInfo:
                  $ref: "#/definitions/SchemaInfo"
              validationRules:
                  $ref: "#/definitions/SchemaValidationRules"
            required:
              - schemaInfo      
      produces:
        - application/json
      responses:
        200:
          description: Schema validation response 
          schema:
             $ref: "#/definitions/Valid"
        404:
          description: Group with given name not found
        500:
          description: Internal server error while trying to validate schema
  /groups/{groupName}/schemas/versions/canRead:
    parameters:
      - in: path
        name: groupName
        description: Group name
        required: true
        type: string
    post:
      tags:
      - "Group"
      operationId: canRead
      description: Checks if given schema can be used for reads subject to compatibility policy in the schema validation rules. 
      consumes:
        - application/json
      parameters:
        - in: body
          name: schemaInfo
          description: Checks if schema can be used to read the data in the stream based on compatibility rules. 
          required: true
          schema:
              $ref: "#/definitions/SchemaInfo"
      produces:
        - application/json
      responses:
        200:
          description: Response to tell whether schema can be used to read existing schemas
          schema:
             $ref: "#/definitions/CanRead"
        404:
          description: Group with given name not found
        500:
          description: Internal server error while checking schema for readability
  /groups/{groupName}/encodings:
    parameters:
      - in: path
        name: groupName
        description: Group name
        required: true
        type: string
    put:
      tags:
      - "Group"
      operationId: getEncodingId
      description: Get an encoding id that uniquely identifies a schema version and codec type pair. 
      consumes:
        - application/json
      parameters:
        - in: body
          name: GetEncodingIdRequest
          description: Get schema corresponding to the version
          required: true
          schema:
            type: object
            properties:
              versionInfo:
                  $ref: "#/definitions/VersionInfo"
              codecType:
                  type: string
            required:
              - versionInfo
              - codecType
      produces:
        - application/json
      responses:
        200:
          description: Found Encoding
          schema:
            $ref: "#/definitions/EncodingId"
        404:
          description: Group with given name or version not found
        412:
          description: Codec type not registered
        500:
          description: Internal server error while getting encoding id
...
definitions:
  ListGroupsResponse:
    type: object
    description: Map of Group names to group properties. For partially created groups, the group properties may be null.   
    properties:
      groups:
        type: object
        additionalProperties:
          $ref: "#/definitions/GroupProperties"
      continuationToken:
        description: Continuation token to identify the position of last group in the response.   
        type: string
    required:
      - continuationToken    
  GroupProperties:
    type: object
    description: Metadata for a group.    
    properties:
      serializationFormat:
          description: serialization format for the group.    
          $ref: "#/definitions/SerializationFormat"
      schemaValidationRules:
          description: Validation rules to apply while registering new schema.    
          $ref: "#/definitions/SchemaValidationRules"
      allowMultipleTypes:
          description: Flag to indicate whether to allow multiple schemas representing distinct objects to be registered in the group.      
          type: boolean
      properties:
        description: User defined Key value strings.     
        type: object
        additionalProperties:
          type: string
          minLength: 0
          maxLength: 40
    required:
      - serializationFormat
      - allowMultipleTypes
      - schemaValidationRules      
...
  SchemaVersionsList:
    type: object
    description: List of schemas with their versions.  
    properties:
      schemas:
        description: List of schemas with their versions.
        type: array
        items:
          $ref: "#/definitions/SchemaWithVersion"
...
  GroupHistory:
    type: object
    properties:
       history:
        type: array
        description: Chronological list of Group History records.
        items:
          $ref: "#/definitions/GroupHistoryRecord"
...